// src/app/interfaces/homework.ts
export interface Homework {
    created_at?: Date | string
    diffDegree?: number
    examDuration?: number
    graded?: number
    id?: number
    lesson_id?: number
    main_exam_id?: number
    maxResult?: number
    questionsNum?: number
    randomQuestions?: number
    subject_id?: number
    subjectTitle?: string
    timeEnd?: Date | string
    timeStart?: Date | string
    start_at?: Date | string
    end_at?: Date | string
    unit_id?: number
    updated_at?: Date | string
    user_id?: number
    unitTitle?: string
    lessonTitle?: string
    duration?: number
    
    max_result?: number
    result?: number
    homework?: any
}